# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  displayName: string;
  avatarUrl: string;
}

export interface ProfileUpdate {
  /**
     * @minLength 1
     * @maxLength 60
     */
  displayName?: string;
  /** @maxLength 500 */
  avatarUrl?: string;
}

export interface Chat {
  id: number;
  participant: User;
  lastMessage: string;
  lastMessageAt: string;
  unreadCount: number;
}

export interface Message {
  id: number;
  chatId: number;
  senderId: number;
  senderName: string;
  content: string;
  createdAt: string;
  isMine: boolean;
}

export interface MessageInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
}

export interface WallPost {
  id: number;
  author: User;
  content: string;
  createdAt: string;
}

export interface WallPostInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  content: string;
}