# lib/api-zod/src/generated/types/chat.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface Chat {
  id: number;
  participant: User;
  lastMessage: string;
  lastMessageAt: string;
  unreadCount: number;
}