# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  /**
     * @minLength 1
     * @maxLength 60
     */
  displayName?: string;
  /** @maxLength 500 */
  avatarUrl?: string;
}