# lib/api-zod/src/generated/types/wallPost.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface WallPost {
  id: number;
  author: User;
  content: string;
  createdAt: string;
}