# lib/api-zod/src/generated/types/wallPostInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WallPostInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  content: string;
}